back/requirements.txt
python-dotenv==1.0.0
fastapi==0.109.*
uvicorn[standard]
python-multipart==0.0.*
requests==2.31.*

aio_pika
authlib
httpx

python-jose[cryptography]
PyJWT==2.6.0
bcrypt

mysqlclient
sqlalchemy

pokemontcgsdk

